+++ includes/libft.h
@@ -43,7 +43,7 @@ typedef struct			s_buff
 t_buff					*ft_buffinit(int size);
 void					ft_buffaddsymb(t_buff *buff, char c);
 void					ft_buffadd(t_buff *buff, const char *str);
-void					ft_buffladd(t_buff *buff, const char *str, int len);
+void					ft_buffladd(t_buff *buff, const char *str, size_t len);
 void					ft_buffdel(t_buff **buff);
 
 /*
